benchdnn: move stream_t related classes into a single file
