+++ tests/benchdnn/dnnl_common.cpp
@@ -356,49 +356,6 @@ void args_t::replace(int arg, const dnn_mem_t *mem) {
     }
 }
 
-stream_staller_t::stream_staller_t(stream_t &stream) {
-#if DNNL_CPU_THREADING_RUNTIME == DNNL_RUNTIME_THREADPOOL
-    auto eng = query_engine(stream);
-    auto eng_kind = query_engine_kind(eng);
-    if (eng_kind != dnnl_cpu) return;
-
-    void *tp_ptr;
-    dnnl_threadpool_interop_stream_get_threadpool(stream, &tp_ptr);
-    auto tp = static_cast<dnnl::threadpool_interop::threadpool_iface *>(tp_ptr);
-
-    // `tp` is not expected to be empty for CPU streams with threadpol runtime.
-    if (!tp) SAFE_V(FAIL);
-
-    // Only relevant for asynchronous threadpool, synchronous will
-    // deadlock.
-    if (tp->get_flags()
-            != dnnl::threadpool_interop::threadpool_iface::ASYNCHRONOUS)
-        return;
-
-    // Each thread from the threadpool should get the task to be stalled.
-    const int num_tasks = tp->get_num_threads();
-
-    // The main thread must be let through, otherwise it deadlocks as
-    // task submission won't happen.
-    std::thread::id main_thr_id = std::this_thread::get_id();
-
-    // Shared future allows to pass all waiting threads at once inside the
-    // palallel call.
-    std::shared_future<void> fut(prom_.get_future());
-
-    tp->parallel_for(num_tasks, [=](int, int) {
-        std::thread::id id_thr = std::this_thread::get_id();
-        if (id_thr != main_thr_id) fut.wait();
-    });
-#endif
-}
-
-void stream_staller_t::release() {
-#if DNNL_CPU_THREADING_RUNTIME == DNNL_RUNTIME_THREADPOOL
-    prom_.set_value();
-#endif
-}
-
 // Unmap before passing the memory to execute
 void execute_unmap_args(
         const args_t &args, std::vector<dnnl_exec_arg_t> &dnnl_args) {
@@ -1759,24 +1716,6 @@ execution_mode_t str2execution_mode(const char *str) {
     BENCHDNN_PRINT(0, "%s", "Error: execution mode value is not recognized.\n");
     SAFE_V(FAIL);
     return execution_mode_t::direct;
-}
-
-stream_t::stream_t(const engine_t &engine, void *interop_obj) {
-#if DNNL_CPU_THREADING_RUNTIME == DNNL_RUNTIME_THREADPOOL
-    if (is_cpu(engine)) {
-        auto tp = static_cast<dnnl::threadpool_interop::threadpool_iface *>(
-                interop_obj);
-        if (tp == nullptr) tp = dnnl::testing::get_threadpool();
-        stream_ = dnnl::threadpool_interop::make_stream(engine, tp);
-        return;
-    }
-#endif
-
-    const bool use_profiling = has_bench_mode_bit(mode_bit_t::perf)
-            && is_gpu(engine) && !is_nvidia_gpu(engine) && !is_amd_gpu(engine);
-    const auto flags = static_cast<dnnl::stream::flags>(
-            stream_kind2stream_flags(stream_kind, use_profiling));
-    stream_ = dnnl::stream(engine, flags);
 }
 
 float reorder_rescale_factor() {
